ackee_veille/requirements.txt
# Core dependencies
anthropic>=0.40.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Web scraping & crawling
selenium>=4.16.0
feedparser>=6.0.10
newspaper3k>=0.2.8

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Scheduling
schedule>=1.2.0
APScheduler>=3.10.0

# Email generation
email-validator>=2.1.0

# API clients
httpx>=0.26.0

# Logging
loguru>=0.7.0

# Date handling
python-dateutil>=2.8.0
